Move minus to thumb
Move minus to thumb in number/symbol layer in the same place as space.
This logically makes some sense as minus often substitutes for space.
This also keeps things logical where underscore (which substitutes for
space similarly is also on the thumb).

This opens up right pinky for transparent quote in number/symbol layer,
but could be later repurposed for something else...

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>