Move minus/underscore onto layer one, enter off
Move minus and (unshifted) underscore onto the first layer at,
respectively, left and right pinky.  This is done to make these very easy
to hit as they come up frequently in both shell interaction and
programming.  This moves escape and enter to the num/symbol layer.  This
is done because escape is less frequently used and makes sense at roughly
its normal position (upper left pinky) on a layer.  I've also gotten very
used to return as a combo (ctrl+j), so I'll see how it feels as a
two-thumb combo.

Fixes #1.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>