Remove duplicate enter
Change the duplication of enter to only be on layer 2.  This doesn't feel
as good, however, it allows enter to work with both ctrl and shift
modifiers which is annoyingly necessary for different chat programs.

Making it so there is only one way of pressing enter makes it better for
consistency. (Enter is always hit the same way and not one way in chat and
one way every else.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>